# [hal][neon] Optimise v_expand for AArch64