Similar to https://github.com/opencv/opencv/pull/19486 we can fuse the two intrinsics
```c++
vmovl_##suffix(vget_high_##suffix());
```
into the equivalent single intrinsic
```c++
vmovl_high_##suffix();
```
for AArch64.

On gcc 10.2 and before, the compiler emits fewer instructions (see https://godbolt.org/z/x9GcY5evj).

### Pull Request Readiness Checklist

See details at https://github.com/opencv/opencv/wiki/How_to_contribute#making-a-good-pull-request

- [x] I agree to contribute to the project under Apache 2 License.
- [x] To the best of my knowledge, the proposed patch is not based on a code under GPL or other license that is incompatible with OpenCV
- [x] The PR is proposed to proper branch
- [ ] There is reference to original bug report and related work
- [ ] There is accuracy test, performance test and test data in opencv_extra repository, if applicable
      Patch to opencv_extra has the same branch name.
- [ ] The feature is well documented and sample code can be built with the project CMake

```
force_builders=linux,docs,ARMv8,ARMv7
```