@jondea 2021-03-25
To benchmark this change, I ran a few perf tests which call `v_expand`:
```sh
./opencv_perf_imgproc --gtest_filter=*cvtColor* --gtest_param_filter=*127*Luv*RG* --perf_min_samples=20 --perf_force_samples=20  --perf_write_validation_results=results.log
```
on a c6gd.4xlarge instance, taking the average of 100 runs before and after the changes. As you can see below, I got a speed up of just under 1%. `v_expand` is used elsewhere too, so there should be improvements elsewhere too.

 test_set | test number | time before changes | time after changes | ratio
-------- | ---------- | ---------------------- | ------------------- | -----
Size_CvtMode_cvtColor8u--cvtColor8u     |   77      |       0.155181    |        0.153927 |  0.99192
Size_CvtMode_cvtColor8u--cvtColor8u     |   78      |       0.161906    |        0.160384 | 0.990604
Size_CvtMode_cvtColor8u--cvtColor8u     |   81       |      0.156524     |       0.155422 | 0.992961
Size_CvtMode_cvtColor8u--cvtColor8u     |   82       |      0.162851      |      0.161776 | 0.993396